restructure(Dependencies): move BlockIndexService and DiscoverService into new DependenciesModule
