+++ demo/api/schema.gql
@@ -71,17 +71,6 @@ type DamFile {
   dependents: [BlockIndexDependency!]!
 }
 
-type BlockIndexDependency {
-  rootId: String!
-  rootGraphqlObjectType: String!
-  rootColumnName: String!
-  blockname: String!
-  jsonPath: String!
-  visible: Boolean!
-  targetGraphqlObjectType: String!
-  targetId: String!
-}
-
 type BuildTemplate {
   id: ID!
   name: String!
@@ -108,6 +97,17 @@ enum JobStatus {
   active
   succeeded
   failed
+}
+
+type BlockIndexDependency {
+  rootId: String!
+  rootGraphqlObjectType: String!
+  rootColumnName: String!
+  blockname: String!
+  jsonPath: String!
+  visible: Boolean!
+  targetGraphqlObjectType: String!
+  targetId: String!
 }
 
 type CronJob {
