+++ packages/api/cms-api/schema.gql
@@ -67,17 +67,6 @@ type DamFile {
   dependents: [BlockIndexDependency!]!
 }
 
-type BlockIndexDependency {
-  rootId: String!
-  rootGraphqlObjectType: String!
-  rootColumnName: String!
-  blockname: String!
-  jsonPath: String!
-  visible: Boolean!
-  targetGraphqlObjectType: String!
-  targetId: String!
-}
-
 type BuildTemplate {
   id: ID!
   name: String!
@@ -104,6 +93,17 @@ enum JobStatus {
   active
   succeeded
   failed
+}
+
+type BlockIndexDependency {
+  rootId: String!
+  rootGraphqlObjectType: String!
+  rootColumnName: String!
+  blockname: String!
+  jsonPath: String!
+  visible: Boolean!
+  targetGraphqlObjectType: String!
+  targetId: String!
 }
 
 type CronJob {
