+++ packages/api/cms-api/src/blocks/block-migrate.service.ts
@@ -3,7 +3,7 @@ import { Connection, EntityManager } from "@mikro-orm/core";
 import { Injectable } from "@nestjs/common";
 import isEqual from "lodash.isequal";
 
-import { DiscoverService } from "./discover.service";
+import { DiscoverService } from "../dependencies/discover.service";
 
 @Injectable()
 export class BlockMigrateService {
