+++ packages/api/cms-api/src/dependencies/dependencies.module.ts
@@ -1,0 +1,12 @@
+import { Global, Module } from "@nestjs/common";
+
+import { BlockIndexService } from "./block-index.service";
+import { DiscoverService } from "./discover.service";
+
+@Global()
+@Module({
+    imports: [],
+    providers: [DiscoverService, BlockIndexService],
+    exports: [DiscoverService, BlockIndexService],
+})
+export class DependenciesModule {}
